Deleted duplicate resources and modified wallaby config to pick up root data sets. Moved class test to import/export test. Moved Data-structure tests to class tests and deleted data-structures because that is not a thing in JS.
