--- src/javascript/basics/arrays/datasets/presidents.js
@@ -1,354 +1,0 @@
-export default [
-  {
-    name: "George Washington",
-    birthday: 1732,
-    death: 1799,
-    tookOffice: "1789-04-30",
-    leftOffice: "1797-03-04",
-    party: "No Party"
-  },
-  {
-    name: "John Adams",
-    birthday: 1735,
-    death: 1826,
-    tookOffice: "1797-03-04",
-    leftOffice: "1801-03-04",
-    party: "Federalist"
-  },
-  {
-    name: "Thomas Jefferson",
-    birthday: 1743,
-    death: 1826,
-    tookOffice: "1801-03-04",
-    leftOffice: "1809-03-04",
-    party: "Democratic-Republican"
-  },
-  {
-    name: "James Madison",
-    birthday: 1751,
-    death: 1836,
-    tookOffice: "1809-03-04",
-    leftOffice: "1817-03-04",
-    party: "Democratic-Republican"
-  },
-  {
-    name: "James Monroe",
-    birthday: 1758,
-    death: 1831,
-    tookOffice: "1817-03-04",
-    leftOffice: "1825-03-04",
-    party: "Democratic-Republican"
-  },
-  {
-    name: "John Quincy Adams",
-    birthday: 1767,
-    death: 1848,
-    tookOffice: "1825-03-04",
-    leftOffice: "1829-03-04",
-    party: "Democratic-Republican"
-  },
-  {
-    name: "Andrew Jackson",
-    birthday: 1767,
-    death: 1845,
-    tookOffice: "1829-03-04",
-    leftOffice: "1837-03-04",
-    party: "Democratic"
-  },
-  {
-    name: "Martin Van Buren",
-    birthday: 1782,
-    death: 1862,
-    tookOffice: "1837-03-04",
-    leftOffice: "1841-03-04",
-    party: "Democratic"
-  },
-  {
-    name: "William Henry Harrison",
-    birthday: 1773,
-    death: 1841,
-    tookOffice: "1841-03-04",
-    leftOffice: "1841-04-04",
-    party: "Whig"
-  },
-  {
-    name: "John Tyler",
-    birthday: 1790,
-    death: 1862,
-    tookOffice: "1841-04-04",
-    leftOffice: "1845-03-04",
-    party: "Whig"
-  },
-  {
-    name: "James K. Polk",
-    birthday: 1795,
-    death: 1849,
-    tookOffice: "1845-03-04",
-    leftOffice: "1849-03-04",
-    party: "Democratic"
-  },
-  {
-    name: "Zachary Taylor",
-    birthday: 1784,
-    death: 1850,
-    tookOffice: "1849-03-04",
-    leftOffice: "1850-07-09",
-    party: "Whig"
-  },
-  {
-    name: "Millard Fillmore",
-    birthday: 1800,
-    death: 1874,
-    tookOffice: "1850-07-09",
-    leftOffice: "1853-03-04",
-    party: "Whig"
-  },
-  {
-    name: "Franklin Pierce",
-    birthday: 1804,
-    death: 1869,
-    tookOffice: "1853-03-04",
-    leftOffice: "1857-03-04",
-    party: "Democratic"
-  },
-  {
-    name: "James Buchanan",
-    birthday: 1791,
-    death: 1868,
-    tookOffice: "1857-03-04",
-    leftOffice: "1861-03-04",
-    party: "Democratic"
-  },
-  {
-    name: "Abraham Lincoln",
-    birthday: 1809,
-    death: 1865,
-    tookOffice: "1861-03-04",
-    leftOffice: "1865-04-15",
-    party: "Republican"
-  },
-  {
-    name: "Andrew Johnson",
-    birthday: 1808,
-    death: 1875,
-    tookOffice: "1865-04-15",
-    leftOffice: "1869-03-04",
-    party: "Democratic"
-  },
-  {
-    name: "Ulysses S. Grant",
-    birthday: 1822,
-    death: 1885,
-    tookOffice: "1869-03-04",
-    leftOffice: "1877-03-04",
-    party: "Republican"
-  },
-  {
-    name: "Rutherford B. Hayes",
-    birthday: 1822,
-    death: 1893,
-    tookOffice: "1877-03-04",
-    leftOffice: "1881-03-04",
-    party: "Republican"
-  },
-  {
-    name: "James A. Garfield",
-    birthday: 1831,
-    death: 1881,
-    tookOffice: "1881-03-04",
-    leftOffice: "1881-09-19",
-    party: "Republican"
-  },
-  {
-    name: "Chester A. Arthur",
-    birthday: 1829,
-    death: 1886,
-    tookOffice: "1881-09-19",
-    leftOffice: "1885-03-04",
-    party: "Republican"
-  },
-  {
-    name: "Grover Cleveland",
-    birthday: 1837,
-    death: 1908,
-    tookOffice: "1885-03-04",
-    leftOffice: "1889-03-04",
-    party: "Democratic"
-  },
-  {
-    name: "Benjamin Harrison",
-    birthday: 1833,
-    death: 1901,
-    tookOffice: "1889-03-04",
-    leftOffice: "1893-03-04",
-    party: "Republican"
-  },
-  {
-    name: "Grover Cleveland",
-    birthday: 1837,
-    death: 1908,
-    tookOffice: "1893-03-04",
-    leftOffice: "1897-03-04",
-    party: "Democratic"
-  },
-  {
-    name: "William McKinley",
-    birthday: 1843,
-    death: 1901,
-    tookOffice: "1897-03-04",
-    leftOffice: "1901-09-14",
-    party: "Republican"
-  },
-  {
-    name: "Theodore Roosevelt",
-    birthday: 1858,
-    death: 1919,
-    tookOffice: "1901-09-14",
-    leftOffice: "1909-03-04",
-    party: "Republican"
-  },
-  {
-    name: "William Howard Taft",
-    birthday: 1857,
-    death: 1930,
-    tookOffice: "1909-03-04",
-    leftOffice: "1913-03-04",
-    party: "Republican"
-  },
-  {
-    name: "Woodrow Wilson",
-    birthday: 1856,
-    death: 1924,
-    tookOffice: "1913-03-04",
-    leftOffice: "1921-03-04",
-    party: "Democratic"
-  },
-  {
-    name: "Warren G. Harding",
-    birthday: 1865,
-    death: 1923,
-    tookOffice: "1921-03-04",
-    leftOffice: "1923-08-02",
-    party: "Republican"
-  },
-  {
-    name: "Calvin Coolidge",
-    birthday: 1872,
-    death: 1933,
-    tookOffice: "1923-08-02",
-    leftOffice: "1929-03-04",
-    party: "Republican"
-  },
-  {
-    name: "Herbert Hoover",
-    birthday: 1874,
-    death: 1964,
-    tookOffice: "1929-03-04",
-    leftOffice: "1933-03-04",
-    party: "Republican"
-  },
-  {
-    name: "Franklin D. Roosevelt",
-    birthday: 1882,
-    death: 1945,
-    tookOffice: "1933-03-04",
-    leftOffice: "1945-04-12",
-    party: "Democratic"
-  },
-  {
-    name: "Harry S. Truman",
-    birthday: 1884,
-    death: 1972,
-    tookOffice: "1945-04-12",
-    leftOffice: "1953-01-20",
-    party: "Democratic"
-  },
-  {
-    name: "Dwight D. Eisenhower",
-    birthday: 1890,
-    death: 1969,
-    tookOffice: "1953-01-20",
-    leftOffice: "1961-01-20",
-    party: "Republican"
-  },
-  {
-    name: "John F. Kennedy",
-    birthday: 1917,
-    death: 1963,
-    tookOffice: "1961-01-20",
-    leftOffice: "1963-11-22",
-    party: "Democratic"
-  },
-  {
-    name: "Lyndon B. Johnson",
-    birthday: 1908,
-    death: 1973,
-    tookOffice: "1963-11-22",
-    leftOffice: "1969-01-20",
-    party: "Democratic"
-  },
-  {
-    name: "Richard Nixon",
-    birthday: 1913,
-    death: 1994,
-    tookOffice: "1969-01-20",
-    leftOffice: "1974-08-09",
-    party: "Republican"
-  },
-  {
-    name: "Gerald Ford",
-    birthday: 1913,
-    death: 2006,
-    tookOffice: "1974-08-09",
-    leftOffice: "1977-01-20",
-    party: "Republican"
-  },
-  {
-    name: "Jimmy Carter",
-    birthday: 1924,
-    death: null,
-    tookOffice: "1977-01-20",
-    leftOffice: "1981-01-20",
-    party: "Democratic"
-  },
-  {
-    name: "Ronald Reagan",
-    birthday: 1911,
-    death: 2004,
-    tookOffice: "1981-01-20",
-    leftOffice: "1989-01-20",
-    party: "Republican"
-  },
-  {
-    name: "George H. W. Bush",
-    birthday: 1924,
-    death: null,
-    tookOffice: "1989-01-20",
-    leftOffice: "1993-01-20",
-    party: "Republican"
-  },
-  {
-    name: "Bill Clinton",
-    birthday: 1946,
-    death: null,
-    tookOffice: "1993-01-20",
-    leftOffice: "2001-01-20",
-    party: "Democratic"
-  },
-  {
-    name: "George W. Bush",
-    birthday: 1946,
-    death: null,
-    tookOffice: "2001-01-20",
-    leftOffice: "2009-01-20",
-    party: "Republican"
-  },
-  {
-    name: "Barack Obama",
-    birthday: 1961,
-    death: null,
-    tookOffice: "2009-01-20",
-    leftOffice: null,
-    party: "Democratic"
-  }
-];
