--- src/javascript/basics/arrays/datasets/tempatures.json
@@ -1,145 +1,0 @@
-{
-  "description": {
-    "title": "Global Land and Ocean Temperature Anomalies, January-December",
-    "units": "Degrees Celsius",
-    "base_period": "1901-2000"
-  },
-  "data": {
-    "1880": "-0.1247",
-    "1881": "-0.0707",
-    "1882": "-0.0710",
-    "1883": "-0.1481",
-    "1884": "-0.2099",
-    "1885": "-0.2220",
-    "1886": "-0.2101",
-    "1887": "-0.2559",
-    "1888": "-0.1541",
-    "1889": "-0.1032",
-    "1890": "-0.3233",
-    "1891": "-0.2552",
-    "1892": "-0.3079",
-    "1893": "-0.3221",
-    "1894": "-0.2828",
-    "1895": "-0.2279",
-    "1896": "-0.0971",
-    "1897": "-0.1232",
-    "1898": "-0.2578",
-    "1899": "-0.1172",
-    "1900": "-0.0704",
-    "1901": "-0.1471",
-    "1902": "-0.2535",
-    "1903": "-0.3442",
-    "1904": "-0.4240",
-    "1905": "-0.2967",
-    "1906": "-0.2208",
-    "1907": "-0.3767",
-    "1908": "-0.4441",
-    "1909": "-0.4332",
-    "1910": "-0.3862",
-    "1911": "-0.4367",
-    "1912": "-0.3318",
-    "1913": "-0.3205",
-    "1914": "-0.1444",
-    "1915": "-0.0747",
-    "1916": "-0.2979",
-    "1917": "-0.3193",
-    "1918": "-0.2118",
-    "1919": "-0.2082",
-    "1920": "-0.2152",
-    "1921": "-0.1517",
-    "1922": "-0.2318",
-    "1923": "-0.2161",
-    "1924": "-0.2510",
-    "1925": "-0.1464",
-    "1926": "-0.0618",
-    "1927": "-0.1506",
-    "1928": "-0.1749",
-    "1929": "-0.2982",
-    "1930": "-0.1016",
-    "1931": "-0.0714",
-    "1932": "-0.1214",
-    "1933": "-0.2481",
-    "1934": "-0.1075",
-    "1935": "-0.1445",
-    "1936": "-0.1173",
-    "1937": "-0.0204",
-    "1938": "-0.0318",
-    "1939": "-0.0157",
-    "1940": "0.0927",
-    "1941": "0.1974",
-    "1942": "0.1549",
-    "1943": "0.1598",
-    "1944": "0.2948",
-    "1945": "0.1754",
-    "1946": "-0.0013",
-    "1947": "-0.0455",
-    "1948": "-0.0471",
-    "1949": "-0.0550",
-    "1950": "-0.1579",
-    "1951": "-0.0095",
-    "1952": "0.0288",
-    "1953": "0.0997",
-    "1954": "-0.1118",
-    "1955": "-0.1305",
-    "1956": "-0.1945",
-    "1957": "0.0538",
-    "1958": "0.1145",
-    "1959": "0.0640",
-    "1960": "0.0252",
-    "1961": "0.0818",
-    "1962": "0.0924",
-    "1963": "0.1100",
-    "1964": "-0.1461",
-    "1965": "-0.0752",
-    "1966": "-0.0204",
-    "1967": "-0.0112",
-    "1968": "-0.0282",
-    "1969": "0.0937",
-    "1970": "0.0383",
-    "1971": "-0.0775",
-    "1972": "0.0280",
-    "1973": "0.1654",
-    "1974": "-0.0698",
-    "1975": "0.0060",
-    "1976": "-0.0769",
-    "1977": "0.1996",
-    "1978": "0.1139",
-    "1979": "0.2288",
-    "1980": "0.2651",
-    "1981": "0.3024",
-    "1982": "0.1836",
-    "1983": "0.3429",
-    "1984": "0.1510",
-    "1985": "0.1357",
-    "1986": "0.2308",
-    "1987": "0.3710",
-    "1988": "0.3770",
-    "1989": "0.2982",
-    "1990": "0.4350",
-    "1991": "0.4079",
-    "1992": "0.2583",
-    "1993": "0.2857",
-    "1994": "0.3420",
-    "1995": "0.4593",
-    "1996": "0.3225",
-    "1997": "0.5185",
-    "1998": "0.6335",
-    "1999": "0.4427",
-    "2000": "0.4255",
-    "2001": "0.5455",
-    "2002": "0.6018",
-    "2003": "0.6145",
-    "2004": "0.5806",
-    "2005": "0.6583",
-    "2006": "0.6139",
-    "2007": "0.6113",
-    "2008": "0.5415",
-    "2009": "0.6354",
-    "2010": "0.7008",
-    "2011": "0.5759",
-    "2012": "0.6219",
-    "2013": "0.6687",
-    "2014": "0.7402",
-    "2015": "0.8990"
-  }
-}
