--- src/javascript/basics/data-structures/README.md
@@ -1,2 +1,0 @@
-# Assignment: Data Structures
-
