+++ src/javascript/basics/objects/classes/classes-and-objects/classes.test.js
@@ -1,34 +1,90 @@
 import Test from 'ava';
-import Robots, { FACTIONS, Robot } from './';
+import * as Assignment from './';
 
-Test('Optimus Prime', t => {
-  const { optimusPrime } = Robots;
-  
-  t.is(optimusPrime.name, 'Optimus Prime');
-  t.is(optimusPrime.faction, FACTIONS.Autobots);
-  t.is(optimusPrime.greet(), `Hi, I'm Optimus Prime`);
+Test('has a Stack class', t => {
+  t.truthy(Assignment.Stack);
 });
 
-Test('Megatron', t => {
-  const { megatron } = Robots;
-  
-  t.is(megatron.name, 'Megatron');
-  t.is(megatron.faction, FACTIONS.Decepticons);
-  t.is(megatron.greet(), `Hi, I'm Megatron`);
+Test('has a Queue class', t => {
+  t.truthy(Assignment.Queue);
 });
 
-Test('Can create bumblebee from Robot', t => {
-  const name = 'BumbleBee';
-  const faction = FACTIONS.Autobots;
-  const POWER = 1000;
-  
-  const bumblebee = new Robot(name, faction);
-  
-  
-  t.is(bumblebee.name, name);
-  t.is(bumblebee.faction, FACTIONS.Autobots);
-  t.is(bumblebee.greet(), `Hi, I'm ${name}`);
-  t.is(bumblebee.power, POWER);
+Test('has a LinkedList class', t => {
+  t.truthy(Assignment.LinkedList);
+});
+
+Test('has a DoubleLinkedList class', t => {
+  t.truthy(Assignment.DoubleLinkedList);
 });
 
 
+Test('has a Tree class', t => {
+  t.truthy(Assignment.Tree);
+});
+
+Test('Stack has push', t => {
+  const myStack = new Assignment.Stack();
+  t.truthy(myStack.push);
+});
+
+Test('Stack has pop', t => {
+  const myStack = new Assignment.Stack();
+  t.truthy(myStack.pop);
+});
+
+Test('Stack can store a value', t => {
+  const stack = new Assignment.Stack();
+  t.is(stack._size, 0);
+  t.deepEqual(stack._storage, {});
+  
+  stack.push(1);
+  t.deepEqual(stack._storage, { '0': 1 });
+});
+
+
+Test('Stack can remove a value', t => {
+  const stack = new Assignment.Stack();
+  t.deepEqual(stack._storage, {});
+  
+  stack.push(1);
+  t.deepEqual(stack._storage, { '0': 1 });
+  const value = stack.pop();
+  t.is(value, 1);
+});
+
+
+Test('Queue has size', t => {
+  const queue = new Assignment.Queue();
+  t.is(queue.size, 0);
+});
+
+Test('Queue has enqueue', t => {
+  const queue = new Assignment.Queue();
+  t.truthy(queue.enqueue);
+});
+
+Test('Queue has dequeue', t => {
+  const queue = new Assignment.Queue();
+  t.truthy(queue.enqueue);
+});
+
+Test('Queue can enqueue an item', t => {
+  const queue = new Assignment.Queue();
+  queue.enqueue(1);
+  t.is(queue.size, 1);
+  t.deepEqual(queue._storage, { '0': 1 });
+});
+
+
+Test('Queue can dequeue an item', t => {
+  const queue = new Assignment.Queue();
+  queue.enqueue(1);
+  t.is(queue.size, 1);
+  t.deepEqual(queue._storage, { '0': 1 });
+  
+  const value = queue.dequeue();
+  t.is(value, 1);
+  t.deepEqual(queue._storage, {});
+  console.log(queue._oldest);
+  console.log(queue._newest);
+});
