+++ src/javascript/basics/packaging-and-building/importing-and-exporting/README.md
@@ -4,14 +4,12 @@
 
 ### Requirements
 
-1. Create a name, age, and isTeacher property in the object literal as the default exports.  Use your name, age, and set the isTeacher to `false`.
-2. Create an additional export for meaningOfLife = 42
-3. Import the `index.js` file in the `imports.js` file.  In the `imports.js` file export the meaning of life from the `index.js` file as the default.
-4. Create a me export in the `imports.js`. Use the default export from the `index.js`.
+1. Make imports.js import your index.js file. Your index.js should hold the bulk of your code, while imports.js is just a medium in which your code is packaged, and then passed to the test cases.
+2. Use the `export` keyword correctly.
 
 
 ### Observations
 
-* Why are modules important? How do I make something public? private?
+* Why are modules important? Using what you did to make your code pass, is it possible to make properties inside your classes inaccessible to the public?
 * What is the difference from the default and exporting a variable?
 * When would you want to export default?  When wouldn't you?
